CLN Refactor Code related to '/crawlLecture' URL
대충 리펙터링한거 같다.

급해서 VueServiceImpl.crawlLecture 메서드 좀 덜한 느낌

일단 끝내긴 끝냄

+++ src/main/java/com/esummary/elearning/repository/user/UserLectureRepository.java
@@ -12,8 +12,8 @@ import com.esummary.elearning.entity.user.UserTask;
 
 @Repository
 public interface UserLectureRepository extends CrudRepository<UserLecture, String>{
-	UserLecture findBySubjectLecture_LectureId(Long lectureId);
+	UserLecture findBySubjectLecture_LectureId(long lectureId);
 	UserLecture findBySubjectLecture(SubjectLecture lecture);
 	
-	UserLecture findByUserSubject_usIdAndSubjectLecture_lectureId(int usId, Long lectureId);
+	UserLecture findByUserSubject_usIdAndSubjectLecture_lectureId(long usId, long lectureId);
 }
